backend/pyproject.toml
[project]
name = "chat-api"
version = "1.0.0"
description = "FastAPI chat application with PostgreSQL and in-memory caching"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "psycopg[binary]>=3.0",
    "SQLAlchemy>=2.0",
    "python-dotenv",
    "pydantic>=2.0",
    "openai>=1.106.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "black",
    "isort",
    "flake8",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
